# e-com/src/main/java/com/ecommerce/dto/CategorySummaryDTO.java
package com.ecommerce.dto;

import com.fasterxml.jackson.annotation.JsonProperty;

/**
 * Lightweight category representation without products to avoid circular references
 */
public class CategorySummaryDTO {
    @JsonProperty("id")
    private Long id;
    
    @JsonProperty("name")
    private String name;
    
    @JsonProperty("description")
    private String description;
    
    // Constructors
    public CategorySummaryDTO() {}
    
    public CategorySummaryDTO(Long id, String name, String description) {
        this.id = id;
        this.name = name;
        this.description = description;
    }
    
    // Getters and Setters
    public Long getId() {
        return id;
    }
    
    public void setId(Long id) {
        this.id = id;
    }
    
    public String getName() {
        return name;
    }
    
    public void setName(String name) {
        this.name = name;
    }
    
    public String getDescription() {
        return description;
    }
    
    public void setDescription(String description) {
        this.description = description;
    }
}